Chatbot(QA)/requirements.txt
streamlit
python-dotenv
Pillow
langchain-core
langchain-openai
langchain_community
langchain_groq
fastapi
uvicorn
langserve
sse_starlette
langchain-huggingface
langchain-chroma
pypdf
bs4
pymupdf
wikipedia
duckdb
pandas
faiss-cpu
pypdf
arxiv
ipykernel
sentence-transformers
huggingface-hub
Pillow
scikit-learn
scipy
torch
tqdm
transformers
langchain-huggingface
langchain
